use argv_split in main as well, avoid wordexp
