+++ mettle/src/argv_split.h
@@ -1,0 +1,6 @@
+#ifndef _ARGV_SPLIT_H_
+#define _ARGV_SPLIT_H_
+
+char ** argv_split(char *args, char **argv, size_t *argc);
+
+#endif
